requirements.txt
# TuvaLLM - Tuvaluan Speech-to-Text Dependencies

# Core ML
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
evaluate>=0.4.0
jiwer>=3.0.0  # WER/CER metrics

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Whisper (optional secondary model)
openai-whisper>=20231117

# LoRA for Whisper
peft>=0.6.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Interactive transcript correction
pygame>=2.5.0  # Audio playback
rich>=13.0.0   # TUI interface

# API server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
